Use Rust FFI: Allow set content type
