+++ src/PhpPact/Consumer/AbstractMessageBuilder.php
@@ -56,4 +56,11 @@ abstract class AbstractMessageBuilder implements BuilderInterface
 
         return $this;
     }
+
+    public function withContentType(?string $contentType): self
+    {
+        $this->message->setContentType($contentType);
+
+        return $this;
+    }
 }
