+++ src/PhpPact/Consumer/Model/Interaction/StatusTrait.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace PhpPact\Consumer\Model\Interaction;
+
+trait StatusTrait
+{
+    private int $status;
+
+    public function getStatus(): int
+    {
+        return $this->status;
+    }
+
+    public function setStatus(int $status): self
+    {
+        $this->status = $status;
+
+        return $this;
+    }
+}
